Refactor FloatValue to DecimalValue

+++ include/radion/interpreter/value.hpp
@@ -5,7 +5,7 @@
 
 enum ValueType {
     // Primitive
-    Int,    Float,
+    Int,    Decimal,
     String, List,
     // defined
     Class,  Func,
